third_party/WebKit/LayoutTests/fast/css/font-face-font-family-descriptor-expected.txt

font-family descriptor in @font-face rule can take only one family. Hence the @font-face rule in this test must be ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test:
Text iii
Must match:
Text iii
Must not match:
Text iii

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById('testText').offsetWidth == document.getElementById('mustMatch').offsetWidth is true
PASS document.getElementById('testText').offsetWidth != document.getElementById('mustNotMatch').offsetWidth is true
